implementation/exercise-1/app/requirements.txt
Flask==3.0.3
web3==6.20.1
eth-account==0.11.2
python-dotenv==1.0.1
